Refactor AddressFormatter import path and add unit tests for address formatting
