+++ src/main/java/uk/gov/justice/laa/portal/landingpage/dto/OfficeDto.java
@@ -6,7 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.justice.laa.portal.landingpage.util.AddressFormatter;
+import uk.gov.justice.laa.portal.landingpage.utils.AddressFormatter;
 
 @Data
 @Builder
