+++ src/main/java/uk/gov/justice/laa/portal/landingpage/model/OfficeModel.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-import uk.gov.justice.laa.portal.landingpage.util.AddressFormatter;
+import uk.gov.justice.laa.portal.landingpage.utils.AddressFormatter;
 
 @Data
 @AllArgsConstructor
@@ -24,7 +24,7 @@ public class OfficeModel {
         private String addressLine2;
         private String city;
         private String postcode;
-        
+
         public String getFormattedAddress() {
             return AddressFormatter.formatAddress(addressLine1, addressLine2, city, postcode);
         }
